chore(ui-ux): move components to its own file
